+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.2.1] - 2021-11-20
 ### Fixed
 - Fixed timeouts after calling `SubGhz::set_sleep`.
 
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 [tweedegolf]: https://github.com/tweedegolf
 [stm32-rs]: https://github.com/stm32-rs
-[Unreleased]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/stm32-rs/stm32wlxx-hal/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/stm32-rs/stm32wlxx-hal/releases/tag/v0.2.0
 [0.1.0]: https://github.com/tweedegolf/stm32wlxx-hal